W24A1 simple graphic
You will create a small graphical Python program. The program will take four integers as input
from the user . Its output will be a picture containing a boat, a sun, and at least one other
entity. The boat and sun in this image will be centered on the positions specified by the user.