Assignment 2
You will create a rudimentary Pictionary game using Python.
Two people will be interacting with your program using the same computer: the drawer and the
guesser .
At the beginning of the game, the drawer will tell the program (while the guesser is not looking)
what the drawing will be. The drawer will also enter in the coordinates of the next point in the
drawing when the guesser asks for it. Use SimpleGraphics to draw a line from the last point to
this new one.