BeatHero
This game, BeatHero, was inspired by the Beat Saber Virtual Reality game. Our game is much simpler,
and you can play it on your computer using your keyboard. Watch the video in the assignment Dropbox
to get a better idea of how the game is supposed to be played (seriously, stop reading here and watch
the video before moving on!)
In our version, when the game starts, the soundtrack starts and beats start falling down your screen, at
first slowly and a few at a time, but as beats move down the zones, they speed up and tests the limits of
your dexterity. Your goal is to score as much as possible before time runs out. Each beat is represented
by an arrow. To score a point, you’ll need to press the arrow key on your keyboard in the direction
noted by the beat arrow before it hits the bottom of the screen. With multiple beats on your screen, you
must always aim for the lowest one.